rename redux folder to store
