+++ src/sagas/talks.saga.js
@@ -3,8 +3,8 @@ import { startSubmit, stopSubmit, reset } from 'redux-form'
 import { push } from 'redux-little-router'
 import compareDesc from 'date-fns/compare_desc'
 
-import store from 'redux/store'
-import { getRouterParam } from 'redux/router'
+import store from 'store/store'
+import { getRouterParam } from 'store/router'
 import talkCrud, { fetchUserTalks } from 'firebase/talks'
 
 function* createTalk(talk) {
