+++ src/screens/components/brand/brand.container.js
@@ -1,5 +1,5 @@
 import { inject } from 'k-ramel/react'
-import { getBaseRoute, getAppTitle } from 'redux/router'
+import { getBaseRoute, getAppTitle } from 'store/router'
 
 import Brand from './brand'
