+++ src/screens/components/event/cfpBlock/cfpBlock.container.js
@@ -1,6 +1,6 @@
 import { inject } from 'k-ramel/react'
 
-import { getCfpState } from 'redux/data/events.selector'
+import { getCfpState } from 'store/data/events.selector'
 import CfpBlock from './cfpBlock'
 
 const mapStore = (store, { eventId }) => {
