+++ src/screens/organizer/event/components/badges/categoryBadge.container.js
@@ -1,6 +1,6 @@
 import { inject } from 'k-ramel/react'
 
-import { getCategory } from 'redux/data/events.selector'
+import { getCategory } from 'store/data/events.selector'
 import Badge from 'components/badge'
 
 const mapStore = (store, { eventId, categoryId }) => {
