+++ src/screens/organizer/event/components/badges/formatBadge.container.js
@@ -1,6 +1,6 @@
 import { inject } from 'k-ramel/react'
 
-import { getFormat } from 'redux/data/events.selector'
+import { getFormat } from 'store/data/events.selector'
 import Badge from 'components/badge'
 
 const mapStore = (store, { eventId, formatId }) => {
