+++ src/screens/speaker/event/submission/talksSelection/talkCardInfo/talkCardInfo.container.js
@@ -1,6 +1,6 @@
 import { inject } from 'k-ramel/react'
 
-import { isSubmitted } from 'redux/data/talks.selector'
+import { isSubmitted } from 'store/data/talks.selector'
 import TalkCardInfo from './talkCardInfo'
 
 const mapStore = (store, { talkId, eventId }) => ({
