+++ src/screens/speaker/speaker.container.js
@@ -2,7 +2,7 @@ import { inject } from 'k-ramel/react'
 import { compose } from 'redux'
 import loader from 'hoc-react-loader/build/core'
 import forRoute from 'hoc-little-router'
-import { protect } from 'redux/auth'
+import { protect } from 'store/auth'
 
 import Speaker from './speaker'
