+++ src/screens/speaker/talk/edit/talkEdit.container.js
@@ -2,7 +2,7 @@ import { compose } from 'redux'
 import { inject } from 'k-ramel/react'
 import forRoute from 'hoc-little-router'
 
-import { getRouterParam } from 'redux/router'
+import { getRouterParam } from 'store/router'
 import loader from 'components/loader'
 import TalkForm from '../components/talkForm'
