+++ src/store/data/ratings.selectors.js
