+++ src/store/data/talks.selector.js
